docker compose for custom rdf loading

+++ openphacts-rdf/download.sh
@@ -1,33 +1,37 @@
-#!/bin/sh
+#!/bin/bash
 set -e
 
 # Inherited from Dockerfile
 #BASE=http://data.openphacts.org/dev/1.5/rdf/
 cd /download
 
-rm -f *.sha1
-# Get *.sha1 first
-echo "Downloading checksums from $BASE"
-wget -e robots=off --no-verbose --recursive --no-directories -A "*.tar*sha1" --no-parent $BASE
+shas=`find -name '*.sha1'`
+if [ "$shas" == "" ] ; then
+  echo "Nothing to download - already finished?"
+  exit 0
+fi
 
+# For every *.sha1 - download from $BASE if not existing
 urls=""
-for sha in *.sha1; do
+for sha in $shas; do
   file=`echo $sha | sed s/.sha1$//`
   if ! [ -f $file ] || ! sha1sum --status -c $sha 2>/dev/null ; then 
+    # Missing or wrong checksum, start again
     rm -f $file
     url="${BASE}${file}"
     urls="$urls $url"
   fi
 done
 
-echo "Downloading RDF datasets to /download"
 
 if [ "$urls" != "" ] ; then
+  echo "Downloading RDF datasets to /download"
   #wget --no-verbose -A "*.tar*" $urls
   for url in $urls; do
       axel --alternate -n8 $url
   done
 fi
+
 
 echo "Verifying checksums"
 sha1sum -c *.sha1
@@ -36,8 +40,10 @@ echo "Extracting to /staging"
 mkdir -p /staging
 cd /staging
 for x in /download/*tar ; do 
+  echo Extracting from $x
   tar xfv $x 
-  # Delete extracted tar file
-  rm $x
+  # Delete extracted tar file so it won't be downloaded+extracted again
+  rm -f $x $x.sha1
 done
 echo "Data download complete"
+
